% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_pkg_repo}
\alias{check_pkg_repo}
\title{Test if package exists.}
\usage{
check_pkg_repo(pkg_name)
}
\arguments{
\item{pkg_name}{}
}
\description{
This function checks if the package exists as a repo in this organisation: https://github.com/cran.
}
\details{
This function is not exported because it does not provide useful information. The function simply checks for if https://github.com/cran/pkg_name gives a 404 error or not. The repo includes packages that are no longer on CRAN (and could be years old!).
}
