% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-docs.R
\name{process_rd_page}
\alias{process_rd_page}
\title{Process an .Rd file into a structured tibble}
\usage{
process_rd_page(rd_path)
}
\description{
This function takes a .Rd filepath and parses the file into a structured tibble.
}
