% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-docs.R
\docType{data}
\name{rd_section_pattern}
\alias{rd_section_pattern}
\title{Rd section patter}
\format{
An object of class \code{character} of length 1.
}
\usage{
rd_section_pattern
}
\description{
Rd section patter
}
\keyword{datasets}
